+++ packages/six-create-account/src/six-create-account.js
@@ -2,26 +2,23 @@ import * as sdk from "@onflow/sdk"
 import * as t from "@onflow/types"
 
 export const TITLE = "Create Account"
-export const DESCRIPTION = "Create an Account on Flow with given code and public keys."
+export const DESCRIPTION = "Create an Account on Flow with public keys."
 export const VERSION = "0.0.0"
-export const HASH = "655d4c6d55b2bb42a208ee9d0793f11499f3f2c3e714cf47dee660ff4185285a"
+export const HASH = "eef2d0494448554177612e63026256258339230cbc6931ded78d6149443c6173"
 export const CODE = 
-`
-transaction(code: String, publicKeys: [String]) {
-    prepare(signer: AuthAccount) {
-        let acct = AuthAccount(payer: signer)
-        for key in publicKeys {
-            acct.addPublicKey(key.decodeHex())
-        }
-        acct.setCode(code.decodeHex())
-    }
+`transaction(publicKeys: [String]) {
+prepare(signer: AuthAccount) {
+let acct = AuthAccount(payer: signer)
+for key in publicKeys {
+acct.addPublicKey(key.decodeHex())
 }
-`
+}
+}`
 
-export const template = ({ proposer, authorization, payer, code = "", publicKeys = [] }) => sdk.pipe([
+export const template = ({ proposer, authorization, payer, publicKeys = [] }) => sdk.pipe([
     sdk.invariant(publicKeys.length > 0, "template({publicKeys}) -- must include one public key when creating an account."),
     sdk.transaction(CODE),
-    sdk.args([sdk.arg(Buffer.from(code, "utf8").toString("hex"), t.String), sdk.arg(publicKeys, t.Array(t.String))]),
+    sdk.args([sdk.arg(publicKeys, t.Array(t.String))]),
     sdk.proposer(proposer),
     sdk.authorizations([authorization]),
     sdk.payer(payer),