+++ packages/six-create-account/CHANGELOG.md
@@ -1,5 +1,11 @@
 ### Unreleased
 
 - YYYY-MM-DD **BREAKING?** -- description
+- 
+### 0.0.1 -- 2020-09-2
+
+- 2020-08-31 -- Update create account code and hash
+- 2020-08-31 -- VSN `@onflow/types` 0.0.2 -> 0.0.4
+- 2020-08-31 -- VSN `@onflow/sdk` 0.0.27 -> 0.0.31
 - 2020-07-13 -- VSN `@onflow/sdk` 0.0.26 -> 0.0.27
 - 2020-07-06 -- Initial Implementation